Refactoring: moving back to show_qr, not functioning yet.
